+++ src/stdash/pages/step_2.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+st.markdown("# STEP 2 ❄️")
+st.sidebar.markdown("# STEP 2 ❄️")
